+++ PHP/JSON-RPC/example.php
@@ -61,7 +61,6 @@ $ssapi->orders(['order_id' => 6], ['name' => 'test order 6', '_group_id' => '333
 echo "REPLACE (w/o answer)\n";
 
 profiling();
-
 profiling(true);