+++ PHP/JSON-RPC/example_local.php
@@ -1,6 +1,6 @@
 <pre>
 <?php
-include(dirname(__FILE__) . '/profiler.php');
+//include(dirname(__FILE__) . '/profiler.php');
 include(dirname(__FILE__) . '/SelectSpecs_v_0_7/include.php');
 
 // we can use custom parameters for connection directly in code (NOT Recomended), please use config.php
@@ -15,57 +15,73 @@ try {
 }
 
 
-//Select orders by query
-    echo "<b>SELECT:</b>\n";
-    $result = $ssapi->orders(['order_id' => 4]);
-    var_dump($result);
+function result_print($result){
+    global $ssapi;
 
-    profiling();
+    if($result){
+        var_dump($result);
+    } elseif($ssapi->getLastError()){
+        echo '<br>ERROR HAPPENNED: <p style="color:red">';
+        var_dump($ssapi->getLastError());
+        echo '</p>';
+    }
+    echo '<hr><p style="color: darkgoldenrod;">execution time is ', $ssapi->getLastTime(),' seconds</p><hr>';
+}
+
+//Select orders by query
+    $result = $ssapi->orders(['order_id' => 4]);
+echo "<h2 style='color: blueviolet;'>SELECT:</h2>";
+result_print($result);
 
 //Remove orders by query
-    echo "<b>Remove:</b>\n";
     $result = $ssapi->orders(['order_id' => 4], NULL, SSAPI_DELETE_DOCUMENT);
-    var_dump($result);
-
-    profiling();
+echo "<h2 style='color: blueviolet;'>REMOVE:</h2>";
+result_print($result);
 
 
 //Replace orders by query
-    echo "<b>REPLACE (with return result flag):</b>\n";
     $result = $ssapi->orders(['order_id' => 3], ['name' => 'test order 3', '_group_id' => '333d3333333333dd33333333'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_RETURN_RESULT);
-    var_dump($result);
-
-    profiling();
+echo "<h2 style='color: blueviolet;'>REPLACE (with return result flag):</h2>";
+result_print($result);
 
 //Insert order
-    echo "<b>INSERT (with return result flag):</b>\n";
     $result = $ssapi->orders(NULL, ['order_id' => 4, 'name' => 'test order 4', '_group_id' => '333d3333333333dd33333333'], SSAPI_RETURN_RESULT);
-    var_dump($result);
-
-    profiling();
+echo "<h2 style='color: blueviolet;'>INSERT (with return result flag):</h2>";
+result_print($result);
 
 //Update orders by query
-    echo "<b>UPDATE (with return result flag):</b>\n";
     $result = $ssapi->orders(['order_id' => 4, '_group_id' => '333d3333333333dd33333333'], ['name' => 'UPDATED order 4 to 2', '_group_id' => '444d4444444444dd44444444'], SSAPI_RETURN_RESULT);
-    var_dump($result);
-
-    profiling();
+echo "<h2 style='color: blueviolet;'>UPDATE (with return result flag):</h2>";
+result_print($result);
 
 //Insert order and don't wait answer from server
     $ssapi->orders(NULL, ['order_id' => 5, 'name' => 'test order 5', '_group_id' => '333d3333333333dd33333333'], SSAPI_NO_WAIT_RESPONSE);
-    echo "INSERT (w/o answer)\n";
-
-    profiling();
+echo "<h2 style='color: blueviolet;'>INSERT (w/o answer):</h2>";
+result_print($result);
 
 //Replace order and don't wait answer from server
     $ssapi->orders(['order_id' => 6], ['name' => 'test order 6', '_group_id' => '333d3333333333dd33333333'], SSAPI_CREATE_IF_NOT_EXIST | SSAPI_NO_WAIT_RESPONSE);
-    echo "REPLACE (w/o answer)\n";
-
-    profiling();
-
-    profiling(true);
+echo "<h2 style='color: blueviolet;'>REPLACE (w/o answer):</h2>";
+result_print($result);
 
 
+if($ssapi->getAllErrors()){
+    echo '<hr><h3>LIST OF ERRORS:</h3> <p style="color:red">';
+    var_dump($ssapi->getAllErrors());
+    echo '</p>';
+} else {
+    echo '<hr><hr><b>NO ERRORS FOUND!</b>';
+}
+if($ssapi->getAllTimes()){
+    echo '<h3>LIST OF QUERIES EXECUTION TIME:</h3> <p style="color:blue;">';
+    var_dump($ssapi->getAllTimes());
+    echo '</p>';
+}
+if($ssapi->getTotalTime()){
+    echo '<hr>Total spended time for all queries: <h2 style="color:olivedrab;">';
+    var_dump($ssapi->getTotalTime());
+    echo '</h2>';
+}
 
 /*
   ERROR CODES: